+++ Changelog.md
@@ -2,6 +2,7 @@ Lilu Changelog
 ==============
 #### v1.4.8
 - Added MacKernelSDK with Xcode 12 compatibility
+- Removed `kern_atomic.hpp` due to MacKernelSDK implementation
 
 #### v1.4.7
 - Added more platform headers for plugin compilation